defining implicits in the companiong object of the given DTOs
