+++ model/src/main/scala/za/co/absa/atum/model/dto/PartitioningSubmitDTO.scala
@@ -16,8 +16,16 @@
 
 package za.co.absa.atum.model.dto
 
+import io.circe.generic.semiauto._
+import io.circe._
+
 case class PartitioningSubmitDTO (
   partitioning: PartitioningDTO,
   parentPartitioning: Option[PartitioningDTO],
   authorIfNew: String
 )
+
+object PartitioningSubmitDTO {
+  implicit val decodePartitioningSubmitDTO: Decoder[PartitioningSubmitDTO] = deriveDecoder
+  implicit val encodePartitioningSubmitDTO: Encoder[PartitioningSubmitDTO] = deriveEncoder
+}
