+++ model/src/main/scala/za/co/absa/atum/model/utils/SerializationUtils.scala
@@ -17,79 +17,108 @@
 package za.co.absa.atum.model.utils
 
 
-import io.circe.{Decoder, Encoder}
-import io.circe.syntax._
-import io.circe.parser._
+//import io.circe.{Decoder, Encoder}
+//import io.circe.syntax._
+//import io.circe.parser._
+//import java.time.ZonedDateTime
+//import java.time.format.DateTimeFormatter
+//import java.util.UUID
+//
+//object SerializationUtils {
+//
+//
+//  val timestampFormat: DateTimeFormatter = DateTimeFormatter.ISO_ZONED_DATE_TIME
+//
+//  implicit val encodeZonedDateTime: Encoder[ZonedDateTime] = Encoder.encodeString.contramap[ZonedDateTime](_.format(timestampFormat))
+//  implicit val decodeZonedDateTime: Decoder[ZonedDateTime] = Decoder.decodeString.emap { str =>
+//    Right(ZonedDateTime.parse(str, timestampFormat))
+//  }
+//
+//  implicit val encodeUUID: Encoder[UUID] = Encoder.encodeString.contramap[UUID](_.toString)
+//  implicit val decodeUUID: Decoder[UUID] = Decoder.decodeString.emap { str =>
+//    Right(UUID.fromString(str))
+//  }
+//
+//  /**
+//   * The method returns arbitrary object as a Json string.
+//   *
+//   * @return A string representing the object in Json format
+//   */
+//  def asJson[T: Encoder](obj: T): String = {
+//    obj.asJson.noSpaces
+//  }
+//
+//  /**
+//   * The method returns arbitrary object as a pretty Json string.
+//   *
+//   * @return A string representing the object in Json format
+//   */
+//  def asJsonPretty[T: Encoder](obj: T): String = {
+//    obj.asJson.spaces2
+//  }
+//
+//  /**
+//   * The method returns arbitrary object parsed from Json string.
+//   *
+//   * @return An object deserialized from the Json string
+//   */
+//  def fromJson[T: Decoder](jsonStr: String): T = {
+//    decode[T](jsonStr) match {
+//      case Right(value) => value
+//      case Left(error) => throw new RuntimeException(s"Failed to decode JSON: $error")
+//    }
+//  }
+//
+//  sealed trait ResultValueType
+//  object ResultValueType {
+//    case object String extends ResultValueType
+//    case object Long extends ResultValueType
+//    case object BigDecimal extends ResultValueType
+//    case object Double extends ResultValueType
+//
+//    implicit val encodeResultValueType: Encoder[ResultValueType] = Encoder.encodeString.contramap {
+//      case ResultValueType.String => "String"
+//      case ResultValueType.Long => "Long"
+//      case ResultValueType.BigDecimal => "BigDecimal"
+//      case ResultValueType.Double => "Double"
+//    }
+//
+//    implicit val decodeResultValueType: Decoder[ResultValueType] = Decoder.decodeString.emap {
+//      case "String" => Right(ResultValueType.String)
+//      case "Long" => Right(ResultValueType.Long)
+//      case "BigDecimal" => Right(ResultValueType.BigDecimal)
+//      case "Double" => Right(ResultValueType.Double)
+//      case other => Left(s"Cannot decode $other as ResultValueType")
+//    }
+//  }
+//
+//}
+
+import io.circe._
+//import za.co.absa.atum.model.dto.MeasureResultDTO.ResultValueType
+
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
 import java.util.UUID
 
 object SerializationUtils {
-
-
-  val timestampFormat: DateTimeFormatter = DateTimeFormatter.ISO_ZONED_DATE_TIME
-
-  implicit val encodeZonedDateTime: Encoder[ZonedDateTime] = Encoder.encodeString.contramap[ZonedDateTime](_.format(timestampFormat))
-  implicit val decodeZonedDateTime: Decoder[ZonedDateTime] = Decoder.decodeString.emap { str =>
-    Right(ZonedDateTime.parse(str, timestampFormat))
-  }
-
-  implicit val encodeUUID: Encoder[UUID] = Encoder.encodeString.contramap[UUID](_.toString)
-  implicit val decodeUUID: Decoder[UUID] = Decoder.decodeString.emap { str =>
-    Right(UUID.fromString(str))
-  }
-
-  /**
-   * The method returns arbitrary object as a Json string.
-   *
-   * @return A string representing the object in Json format
-   */
-  def asJson[T: Encoder](obj: T): String = {
-    obj.asJson.noSpaces
-  }
-
-  /**
-   * The method returns arbitrary object as a pretty Json string.
-   *
-   * @return A string representing the object in Json format
-   */
-  def asJsonPretty[T: Encoder](obj: T): String = {
-    obj.asJson.spaces2
-  }
-
-  /**
-   * The method returns arbitrary object parsed from Json string.
-   *
-   * @return An object deserialized from the Json string
-   */
-  def fromJson[T: Decoder](jsonStr: String): T = {
-    decode[T](jsonStr) match {
-      case Right(value) => value
-      case Left(error) => throw new RuntimeException(s"Failed to decode JSON: $error")
-    }
-  }
-
-  sealed trait ResultValueType
-  object ResultValueType {
-    case object String extends ResultValueType
-    case object Long extends ResultValueType
-    case object BigDecimal extends ResultValueType
-    case object Double extends ResultValueType
-
-    implicit val encodeResultValueType: Encoder[ResultValueType] = Encoder.encodeString.contramap {
-      case ResultValueType.String => "String"
-      case ResultValueType.Long => "Long"
-      case ResultValueType.BigDecimal => "BigDecimal"
-      case ResultValueType.Double => "Double"
-    }
-
-    implicit val decodeResultValueType: Decoder[ResultValueType] = Decoder.decodeString.emap {
-      case "String" => Right(ResultValueType.String)
-      case "Long" => Right(ResultValueType.Long)
-      case "BigDecimal" => Right(ResultValueType.BigDecimal)
-      case "Double" => Right(ResultValueType.Double)
-      case other => Left(s"Cannot decode $other as ResultValueType")
-    }
-  }
+//  // Timestamp format
+//  val timestampFormat: DateTimeFormatter = DateTimeFormatter.ISO_ZONED_DATE_TIME
+//
+//  // Implicit encoders and decoders for ZonedDateTime
+//  implicit val encodeZonedDateTime: Encoder[ZonedDateTime] = Encoder.encodeString.contramap[ZonedDateTime](_.format(timestampFormat))
+//  implicit val decodeZonedDateTime: Decoder[ZonedDateTime] = Decoder.decodeString.emap { str =>
+//    Right(ZonedDateTime.parse(str, timestampFormat))
+//  }
+//
+//  // Implicit encoders and decoders for UUID
+//  implicit val encodeUUID: Encoder[UUID] = Encoder.encodeString.contramap[UUID](_.toString)
+//  implicit val decodeUUID: Decoder[UUID] = Decoder.decodeString.emap { str =>
+//    Right(UUID.fromString(str))
+//  }
+//
+//  // Implicit encoders and decoders for Option[String]
+//  implicit val decodeOptionString: Decoder[Option[String]] = Decoder.decodeOption[String]
+//  implicit val encodeOptionString: Encoder[Option[String]] = Encoder.encodeOption[String]
 
 }
